Move DynamicFilterData to table filter functions
