+++ src/execution/operator/order/physical_top_n.cpp
@@ -6,7 +6,7 @@
 #include "duckdb/execution/expression_executor.hpp"
 #include "duckdb/function/create_sort_key.hpp"
 #include "duckdb/storage/data_table.hpp"
-#include "duckdb/planner/filter/dynamic_filter.hpp"
+#include "duckdb/planner/filter/table_filter_function_helpers.hpp"
 
 namespace duckdb {
